apps/shifts/tests/models/test_application.py
from django.test import TestCase

from apps.shifts.factories import ApplicationFactory, ShiftFactory
from apps.shifts.models import ApplicationStateEnum


class ApplicationTest(TestCase):
    def setUp(self):
        super(ApplicationTest, self).setUp()
        self.shift = ShiftFactory.create()

    def test_approve(self):
        # New applications
        first_application = ApplicationFactory.create(shift=self.shift)
        second_application = ApplicationFactory.create(shift=self.shift)

        another_shift_application = ApplicationFactory.create()

        # Cancelled applications
        cancelled_application = ApplicationFactory.create(
            shift=self.shift,
            state=ApplicationStateEnum.CANCELLED)

        first_application.approve()
        first_application.save()

        first_application.refresh_from_db()
        self.assertEqual(
            first_application.state, ApplicationStateEnum.APPROVED)

        second_application.refresh_from_db()
        self.assertEqual(
            second_application.state, ApplicationStateEnum.REJECTED)

        cancelled_application.refresh_from_db()
        self.assertEqual(
            cancelled_application.state, ApplicationStateEnum.CANCELLED)

        another_shift_application.refresh_from_db()
        self.assertEqual(
            another_shift_application.state, ApplicationStateEnum.NEW)
